+++ src/main/java/ru/sporting/goods/rental/services/TypeOfItemService.java
@@ -2,6 +2,7 @@ package ru.sporting.goods.rental.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.sporting.goods.rental.exceptions.RecordNotFound;
 import ru.sporting.goods.rental.exceptions.TypeOfItemNotFoundException;
 import ru.sporting.goods.rental.entities.TypeOfItem;
 import ru.sporting.goods.rental.repositories.TypeOfItemRepository;
@@ -21,7 +22,7 @@ public class TypeOfItemService {
     //Получение одного продукта по id
     public TypeOfItem getOne(Long id){
         return typeOfItemRepository.findById(id)
-                .orElseThrow(() -> new TypeOfItemNotFoundException(id));
+                .orElseThrow(() -> new RecordNotFound(id));
     }
 
     //Добавление нового типа товара
@@ -29,22 +30,26 @@ public class TypeOfItemService {
         typeOfItemRepository.save(product);
     }
 
+    //Метод для проверки наличия типа товара в базе
+    private boolean checkRecordToBase(Long id){
+        return typeOfItemRepository.existsById(id);
+    }
+
     //Редактирование типа товара
     public void updateTypeProduct(TypeOfItem product){
-        boolean exists = typeOfItemRepository.existsById(product.getId());
-        if (exists){
+        if (checkRecordToBase(product.getId())){
             typeOfItemRepository.save(product);
         } else{
-            throw new TypeOfItemNotFoundException(product.getId());
+            throw new RecordNotFound(product.getId());
         }
     }
 
     //Удаление типа продукта по id
     public void deleteProductTypeById(Long id) {
-        try {
+        if (checkRecordToBase(id)){
             typeOfItemRepository.deleteById(id);
-        } catch (IllegalArgumentException e) {
-            throw new TypeOfItemNotFoundException(id);
+        } else{
+            throw new RecordNotFound(id);
         }
     }
 }
